download as zip, use vite
